# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teams",
        "description": "Team Actions Controller",
        "version": "1.0.0"
    },
    "paths": {
        "/counter": {
            "post": {
                "tags": [
                    "Counters"
                ],
                "summary": "Store a counter",
                "operationId": "a32c478be57986eeb10cd86c53a2ea39",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Dev team",
                                    "team_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/counter/{counter}": {
            "delete": {
                "tags": [
                    "Counters"
                ],
                "summary": "Delete a counter",
                "operationId": "c64a4b361b0abfb10def106f6a2c2b1e",
                "parameters": [
                    {
                        "name": "counter",
                        "in": "path",
                        "description": "Counter id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content. Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/employee/{employee}/counter/{counter}": {
            "post": {
                "tags": [
                    "Counters"
                ],
                "summary": "Increment specific step counter for an employee and a common step counter for a team",
                "operationId": "df5fbf171a97031fb8c7f01f5cd92f44",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "Employee id",
                        "required": true
                    },
                    {
                        "name": "counter",
                        "in": "path",
                        "description": "Counter id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "An employee and a counter have different teams"
                    }
                }
            }
        },
        "/team": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Store a team",
                "operationId": "0a13237ce291eafb4272c524322885f2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Dev team"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/team/{team}": {
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Delete a team",
                "operationId": "6c8af990df20ea9cd3297f3f14fd0eb1",
                "parameters": [
                    {
                        "name": "team",
                        "in": "path",
                        "description": "Team id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content. Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/team/{team}/total/steps": {
            "get": {
                "tags": [
                    "Teams Counters"
                ],
                "summary": "Get the current total steps taken by a team",
                "operationId": "3d95e4b6699972701505a9a890b28316",
                "parameters": [
                    {
                        "name": "team",
                        "in": "path",
                        "description": "Team id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/team/{team}/employees/steps": {
            "get": {
                "tags": [
                    "Teams Counters"
                ],
                "summary": "Get how much each team member have walked. The list of team employees with counters",
                "operationId": "54d4b02b171cdd16c677204d4e1a4aff",
                "parameters": [
                    {
                        "name": "team",
                        "in": "path",
                        "description": "Team id",
                        "required": true
                    },
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "Employees per page"
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Current page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}